+++ src/messages/layouts/MessageLayout.hpp
@@ -37,6 +37,7 @@ public:
     ~MessageLayout();
 
     const Message *getMessage();
+    const MessagePtr &getMessagePtr() const;
 
     int getHeight() const;
 
@@ -62,6 +63,8 @@ public:
 
     // Misc
     bool isDisabled() const;
+    bool isReplyable() const;
+    void setRenderReplies(bool render);
 
 private:
     // variables
@@ -69,6 +72,7 @@ private:
     std::shared_ptr<MessageLayoutContainer> container_;
     std::shared_ptr<QPixmap> buffer_{};
     bool bufferValid_ = false;
+    bool renderReplies_ = true;
 
     int height_ = 0;